+++ lib/engine/index.ts
@@ -1,2 +1,6 @@
 export * from './types';
 export * from './city-pack';
+export * from './scoring';
+export * from './state';
+export * from './random-events';
+export * from './interruption';